integrationtests fixed #1161
- only build zip file if seomthing is defined
- fail if zip file is empty
- moved some funcs into zipbuilder.go
- small improvements
