+++ sechub-cli/src/mercedes-benz.com/sechub/util/fileHelpers.go
@@ -86,7 +86,7 @@ func VerifyDirectoryExists(directory string) bool {
 	return true
 }
 
-// FindNewestMatchingFileInDir - used e.g. for finding the latest report file
+// GetFileSize - return file size in bytes
 func GetFileSize(filepath string) int64 {
 	fileinfo, err := os.Stat(filepath)
 	HandleIOError(err)
